backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
sentence-transformers
faiss-cpu
unstructured
python-multipart
pydantic
loguru
orjson
tenacity
httpx
gunicorn
python-dotenv
aiohttp
pytest